test/wpt/wpt-alt/is-where-pseudo-classes.html

<!DOCTYPE html>
<title>:is() combined with pseudo-classes</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/selectors-4/#matches">
<link rel="help" href="https://drafts.csswg.org/selectors/#useraction-pseudos">
<style>
  button {
    color: black;
  }
  /* Selects #a, #c */
  :is(main :where(main #a), #c:nth-child(odd), #d):is(:enabled) {
    color: green;
  }
  /* Selects #b, #d, #f */
  button:is(:nth-child(even), span #e):is(:enabled, :where(:disabled)) {
    color: blue;
  }
</style>
<main>
  <button id=a>A</button>
  <button id=b>B</button>
  <button id=c>C</button>
  <button id=d disabled>D</button>
  <button id=e disabled>E</button>
  <button id=f disabled>F</button>
</main>
<script>
  test(function() {
    const colorGreen =
      ':is(main :where(main #a), #c:nth-child(odd), #d):is(:enabled)';
    const colorBlue =
      'button:is(:nth-child(even), span #e):is(:enabled, :where(:disabled))';
    assert_true(a.matches(colorGreen));
    assert_true(b.matches(colorBlue));
    assert_true(c.matches(colorGreen));
    assert_true(d.matches(colorBlue));
    assert_false(e.matches(colorGreen));
    assert_false(e.matches(colorBlue));
    assert_true(f.matches(colorBlue));
  }, ':is() combined with pseudo-classes');
</script>
